delete comments rename some files delete some files
