+++ ASP.NETCoreWebApplication/ASP.NETCoreWebApplication.csproj
@@ -53,7 +53,6 @@
     </ItemGroup>
     <ItemGroup>
       <Folder Include="frontend\src\components\pages" />
-      <Folder Include="Models\MySQLSchema" />
       <Folder Include="Tests" />
     </ItemGroup>
     <ItemGroup>
