+++ ASP.NETCoreWebApplication/Infrastructure/InvalidLocationException.cs
@@ -3,10 +3,5 @@
 namespace ASP.NETCoreWebApplication.Utils
 {
     public class InvalidLocationException : Exception
-    {
-        public InvalidLocationException()
-        {
-            
-        }
-    }
+    { }
 }
