+++ ASP.NETCoreWebApplication/Infrastructure/Logger.cs
