+++ ASP.NETCoreWebApplication/Models/DataSources/AlioLt.cs
@@ -5,5 +5,4 @@
         private static string href = "https://www.alio.lt/paieska/?category_id=1373&search_block=1";
     }
     
-    
 }
