--- ASP.NETCoreWebApplication/Models/DataSources/Amazon.cs
@@ -1,7 +1,0 @@
-﻿namespace ASP.NETCoreWebApplication.Models.DataSources
-{
-    public class Amazon
-    {
-        
-    }
-}
