--- ASP.NETCoreWebApplication/Models/DataSources/CurrencyRateScrapper.cs
@@ -1,7 +1,0 @@
-﻿namespace ASP.NETCoreWebApplication.Models.DataSources
-{
-    public class CurrencyRateScrapper
-    {
-        
-    }
-}
