--- ASP.NETCoreWebApplication/Models/DataSources/Ebay.cs
@@ -1,7 +1,0 @@
-﻿namespace ASP.NETCoreWebApplication.Models.DataSources
-{
-    public class Ebay
-    {
-        
-    }
-}
