+++ ASP.NETCoreWebApplication/Models/DataSources/Senukai.cs
@@ -1,6 +1,6 @@
 ﻿namespace ASP.NETCoreWebApplication.Models.DataSources
 {
-    public class Senukai
+    public class enukai
     {
         
     }
